third_party/blink/web_tests/platform/generic/storage/indexeddb/mozilla/key-requirements-put-null-key-expected.txt

Test IndexedDB's behavior put()ing with null key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "key-requirements-put-null-key.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('bar');
Expecting exception from objectStore.put({}, null);
PASS Exception was thrown.
PASS code is 0
PASS ename is 'DataError'
Exception message: Failed to execute 'put' on 'IDBObjectStore': The parameter is not a valid key.
PASS successfullyParsed is true

TEST COMPLETE
